chore: move logger interface to log/
